Improved affected stations accuracy
In this new version, I've added the `validate_stations()` function that ensures that when we look for affected stations based on our manually maintained mapping table, we only consider those stations listed in there that actually pertain to the metro line recorded in the input data. E.g. if a status message for both M1 and M2 lists "Kastrup" as affected but the line pertains to the M1 line, then "Kastrup" will be removed from the list of affected stations on that row.