Adapted scraper to new metro website
Since they changed the HTML code, the scraper was not working. I've now adjusted it to the new structure of the website, and it performs fine when all lines have normal operations. However, I have no idea what the website looks like when there are issues, so I may need to slightly adjust the scraper when different status messages start appearing.