Added log export if HTML extract fails
As the metro's website has been updated, I don't yet know what it looks like if there isn't normal service. Therefore, I've enabled writing the HTML code to a local file under the "logs" folder to capture what the web page looks like if I can't extract the operational status correctly. This should make it easier to fix the stuff when I get the time to work on it.