Optimized df memory for streamlit app
Previously, data frames were quite big because `dtype` optimization was not performed. Now, I have used downcasting for floats and integers, and have converted strings to categories. This has led to a massive reduction of memory usage, between 80-96% for the most relevant data frames. I hope this reduces resource use on the streamlit platform and makes the app speedier. As a next step, I might need to optimize storage in the data frames created in the `app.py` script itself.