Added notebook for troubleshooting web scraper
This notebook can be used to help troubleshoot the web scraping method, particularly the extraction of the lines and status messages. It does not require `selenium` as it expects the user to copy-paste the HTML code from the web browser, and thus allows for remote troubleshooting and quickly patching the code (if needed).